fix number conversion

+++ WebVella.ERP.Storage.Mongo/Impl/MongoTypeConvertor.cs
@@ -18,17 +18,17 @@ namespace WebVella.ERP.Storage.Mongo
                 throw new ArgumentException("'number' argument is not a number type");
 
             if (number is sbyte)
-                return Convert.ToInt64(((sbyte)number)) * PRECISION_OFFSET;
+                return Convert.ToInt64( (long)((sbyte)number)) * PRECISION_OFFSET;
             else if (number is byte)
-                return Convert.ToInt64(((byte)number)) * PRECISION_OFFSET;
+                return Convert.ToInt64((long)((byte)number)) * PRECISION_OFFSET;
             else if (number is short)
-                return Convert.ToInt64(((short)number)) * PRECISION_OFFSET;
+                return Convert.ToInt64((long)((short)number)) * PRECISION_OFFSET;
             else if (number is ushort)
-                return Convert.ToInt64(((ushort)number)) * PRECISION_OFFSET;
+                return Convert.ToInt64((long)((ushort)number)) * PRECISION_OFFSET;
             else if (number is int)
-                return Convert.ToInt64(((int)number)) * PRECISION_OFFSET;
+                return Convert.ToInt64((long)((int)number)) * PRECISION_OFFSET;
             else if (number is uint)
-                return Convert.ToInt64(((uint)number)) * PRECISION_OFFSET;
+                return Convert.ToInt64((long)((uint)number)) * PRECISION_OFFSET;
             else if (number is long)
                 return Convert.ToInt64(((long)number)) * PRECISION_OFFSET;
             else if (number is ulong)
